Extract UserRepository Interface
Converted UserRepository into interface and created its implementation to make testing easier and to follow Interface Segregation Principle.
