+++ app/src/main/java/com/skymeet/videoConference/di/AppModule.java
@@ -1,10 +1,15 @@
 package com.skymeet.videoConference.di;
 
+import androidx.annotation.NonNull;
+
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.firestore.FirebaseFirestore;
+import com.skymeet.videoConference.data.repository.UserRepository;
+import com.skymeet.videoConference.data.repository.UserRepositoryImpl;
 
 import javax.inject.Singleton;
 
+import dagger.Binds;
 import dagger.Module;
 import dagger.Provides;
 import dagger.hilt.InstallIn;
@@ -12,17 +17,23 @@ import dagger.hilt.components.SingletonComponent;
 
 @Module
 @InstallIn(SingletonComponent.class)
-public class AppModule {
+public abstract class AppModule {
 
+    @NonNull
     @Provides
     @Singleton
-    public FirebaseAuth providesFirebaseAuth() {
+    public static FirebaseAuth providesFirebaseAuth() {
         return FirebaseAuth.getInstance();
     }
 
+    @NonNull
     @Provides
     @Singleton
-    public FirebaseFirestore providesFirebaseFirestore() {
+    public static FirebaseFirestore providesFirebaseFirestore() {
         return FirebaseFirestore.getInstance();
     }
+
+    @Binds
+    @Singleton
+    public abstract UserRepository providesUserRepository(UserRepositoryImpl userRepository);
 }
